restructuring: removing ModelInfo which is useless, to log instead the full ModelConfig object. Also moved some classes around as they were not in the right place
