+++ src/lighteval/pipeline.py
@@ -176,9 +176,7 @@ class Pipeline:
         self.accelerator, self.parallel_context = self._init_parallelism_manager()
         self.model = self._init_model(model_config, model)
 
-        generation_parameters = model_config.generation_parameters.model_dump() if model_config else {}
-
-        self.evaluation_tracker.general_config_logger.log_model_info(generation_parameters, self.model.model_info)
+        self.evaluation_tracker.general_config_logger.log_model_info(model_config=self.model.config)
 
         self._init_random_seeds()
         self._init_tasks_and_requests(tasks=tasks)
@@ -293,7 +291,6 @@ class Pipeline:
             num_fewshot_seeds=self.pipeline_parameters.num_fewshot_seeds,
             max_samples=self.pipeline_parameters.max_samples,
             job_id=str(self.pipeline_parameters.job_id),
-            config=self.model_config,
         )
 
         if self.pipeline_parameters.load_responses_from_details_date_id:
